Nothing happens anymore on import of glia module.
It turned out to be an annoying design, restricting options and causing 
issues for a minor gain in convenience.

Instead `import glia, glia.library` or `import glia; 
glia.load_library()` can be used to the same end.

Lastly, just `import glia` can suffice as well because any action that 
requires the glia library is wrapped in `_requires_library`, loading the 
library before executing the function